refactor: abstract out logic and component
