+++ src/components/room-list/index.tsx
@@ -1,0 +1,4 @@
+import { AddRoomDialog } from "./AddRoomDialog";
+import { RoomListCard } from "./RoomListCard";
+
+export { AddRoomDialog, RoomListCard };
