+++ src/hooks/room-list/get-room-id.ts
@@ -2,8 +2,8 @@ import { RoomBasicInfoResponse } from "@/proto-generated/nori/v0/room/room_basic
 
 
 export const getRoomId = (room: RoomBasicInfoResponse): bigint => {
-    if (!room.roomId) {
-      throw new Error("Room ID is undefined");
-    }
-    return room.roomId.id;
-  };
+  if (!room.roomId) {
+    throw new Error("Room ID is undefined");
+  }
+  return room.roomId.id;
+};
