+++ src/hooks/room-list/get-room-name.ts
@@ -2,6 +2,6 @@ import { RoomBasicInfoResponse } from "@/proto-generated/nori/v0/room/room_basic
 
 
 export const getRoomName = (room: RoomBasicInfoResponse): string => {
-    return room.name.case === "sharedName" ? room.name.value :
-      room.name.case === "customName" ? room.name.value : "";
+  return room.name.case === "sharedName" ? room.name.value :
+    room.name.case === "customName" ? room.name.value : "";
 };
